requirements.txt
# =============================================================================
# requirements.txt — DSDBA Pinned Dependencies
# DSDBA-SRS-2026-002 v2.1 | NFR-Security: exact versions required
# Generated from: Google Colab Free Tier (Python 3.10, CUDA 11.8 / T4)
# Serving environment: HuggingFace Spaces (CPU-only). README frontmatter MUST include:
#   python_version: "3.10"
# Otherwise the default Python 3.13 image cannot install torch==2.1.0 (no cp313 wheels).
# =============================================================================

# ── Core ML Framework ─────────────────────────────────────────────────────────
torch==2.1.0                     # FR-CV-001: PyTorch 2.x
torchvision==0.16.0              # FR-CV-001: EfficientNet-B4 pretrained weights
torchaudio==2.1.0                # FR-AUD-001: audio I/O (MP3/OGG optional formats)

# ── Audio DSP ─────────────────────────────────────────────────────────────────
librosa==0.10.1                  # FR-AUD-006: Mel spectrogram extraction
soundfile==0.12.1                # FR-AUD-001: WAV/FLAC read support
numpy==1.24.4                    # FR-AUD-007: numerical operations
setuptools==69.5.1               # pkg_resources (librosa lazy import; missing on minimal venvs)

# ── Computer Vision / XAI ─────────────────────────────────────────────────────
grad-cam==1.5.0                  # FR-CV-011: jacobgil/pytorch-grad-cam
Pillow==10.0.1                   # FR-CV-012: PNG image generation for heatmaps
matplotlib==3.7.3                # FR-CV-012: jet colormap rendering
scikit-learn==1.3.0              # FR-CV-008: EER / AUC-ROC evaluation metrics

# ── ONNX Runtime ──────────────────────────────────────────────────────────────
onnx==1.14.1                     # FR-DEP-010: ONNX model format
# 1.16.x wheels hit "cannot enable executable stack" on glibc 2.41+ (HF Spaces).
onnxruntime==1.23.2              # FR-DEP-010: CPU inference — CPUExecutionProvider

# ── NLP / API ─────────────────────────────────────────────────────────────────
openai==1.3.0                    # FR-NLP-002: Qwen 2.5 API client (OpenAI-compatible)
httpx==0.25.0                    # FR-NLP-002: async HTTP client for API calls
aiohttp==3.8.6                   # FR-NLP-006: async invocation isolation

# ── Deployment / UI ───────────────────────────────────────────────────────────
# Gradio is installed by Hugging Face Spaces (gradio[oauth]==sdk_version from README).
# Do not pin gradio here — it conflicts with the Space builder's pip install line.
# Local dev: pip install "gradio[oauth]==4.44.0" (match README sdk_version).
pydantic==2.4.2                  # config.yaml validation (src/utils/config.py)
python-dotenv==1.0.0             # .env.example support for local development

# ── HuggingFace Ecosystem ─────────────────────────────────────────────────────
huggingface-hub==0.19.4          # FR-CV-007: Model Hub upload/download
datasets==2.14.6                 # FoR dataset access via HuggingFace datasets

# ── Configuration & Utilities ─────────────────────────────────────────────────
PyYAML==6.0.1                    # config.yaml loading
pathlib2==2.3.7.post1            # Path operations (Python 3.10 compatibility)

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.3                    # All test_*.py files
pytest-asyncio==0.21.1           # Async test support (test_nlp.py)
pytest-timeout==2.2.0            # Latency assertion enforcement

# ── Code Quality (development only) ──────────────────────────────────────────
black==23.9.1                    # R.E.F.A.C.T.: code formatting
mypy==1.6.1                      # R.E.F.A.C.T.: strict type checking
flake8==6.1.0                    # R.E.F.A.C.T.: style compliance
pydub